fix: Currency decimals (#281)
* fix: Show tokens amoutn instead of count for likes/dislikes/flags

* fix: Use base PHOTON istead on UPHOTON, add tools for currencies

* fix: Use atomic photon for actions

* fix: after merge

* fix: Change Filter label to "micro", remove unused file

* chore: rollback merge

* fix: Enhance currencies manipulations using cosmjs/math

* fix: FilterIPhoton and InputPhoton, enhance Decimal usages

* fix:Replace function usage

* fix: lint:fix

* fix: Adjust PostActions

* fix: useChain functions parameters

* fix: Simplify Decimal usages

* fix: bigint instead of number in DB schema, enchance InputPhoton

* fix: default bigint value using sql namespace in schema.ts

* fix: Replace quantity/minQuantity bigint by string

* fix: Drizzle sql operation

+++ packages/api-main/src/posts/flag.ts
@@ -25,7 +25,7 @@ const statementAddFlagToPost = getDatabase()
     .update(FeedTable)
     .set({
         flags: sql`${FeedTable.flags} + 1`,
-        flags_burnt: sql`${FeedTable.flags_burnt} + ${sql.placeholder('quantity')}`,
+        flags_burnt: sql`(${FeedTable.flags_burnt})::int + ${sql.placeholder('quantity')}`,
     })
     .where(eq(FeedTable.hash, sql.placeholder('post_hash')))
     .prepare('stmnt_add_flag_count_to_post');
